Extract AssetSelectionControl UserControl from DeployRuntimeView
DeployRuntimeView's picker UI (source folder, target framework combo,
ref-assemblies checkbox, assembly ListView) is now a standalone
AssetSelectionControl UserControl, exposing GetCheckedAssemblies(),
FindFirstWithRefAssemblies(), SelectedSourceTarget, HasAssemblies, and an
AvailabilityChanged event. DeployRuntimeView now hosts it (DI-injected,
Transient) alongside its existing 'Copy' action panel; behavior of the
direct-copy tab is unchanged. This lets a future dialog host the same
picker UI with a different action panel ('Create package installer...').
