+++ src/DevTools.RuntimeDeploy/Views/AssetSelectionControl.cs
@@ -1,0 +1,253 @@
+using DevTools.RuntimeDeploy.Engine.Domain;
+using DevTools.RuntimeDeploy.Infrastructure;
+using Microsoft.Extensions.Logging;
+using static DevTools.RuntimeDeploy.Engine.Domain.BuildArtefactsScanner;
+
+namespace DevTools.RuntimeDeploy.Views;
+
+/// <summary>
+///  Reusable picker UI for the WinForms runtime artefacts source folder, the
+///  target framework, and the set of assemblies to include. Extracted out of
+///  <see cref="DeployRuntimeView"/> so it can also be hosted by
+///  <see cref="CreateRuntimePatcherForm"/> - both share this selection UI but
+///  present a different action panel next to it ("Copy" vs. "Create package
+///  installer...").
+/// </summary>
+public partial class AssetSelectionControl : UserControl
+{
+    private readonly Control[] _controlsForEnablingHandling;
+    private BuildArtefactsScanner? _gitHubRepoManager;
+    private RuntimeDeploySettingsService? _settings;
+    private RuntimeDeployStatusService? _statusService;
+    private ILogger<AssetSelectionControl>? _logger;
+
+    private const string ACCESSIBILITY = "Accessibility";
+    private const string MICROSOFT_VISUALBASIC = "Microsoft.VisualBasic";
+    private const string MICROSOFT_VISUALBASIC_FACADE = "Microsoft.VisualBasic.Facade";
+    private const string MICROSOFT_VISUALBASIC_FORMS = "Microsoft.VisualBasic.Forms";
+    private const string MICROSOFT_PRIVATE_WINFORMS = "Microsoft.Private.Winforms";
+    private const string SYSTEM_DESIGN_FACADE = "System.Design.Facade";
+    private const string SYSTEM_DRAWING_COMMON = "System.Drawing.Common";
+    private const string SYSTEM_DRAWING_DESIGN_FACADE = "System.Drawing.Design.Facade";
+    private const string SYSTEM_DRAWING_FACADE = "System.Drawing.Facade";
+    private const string SYSTEM_PRIVATE_WINDOWS_CORE = "System.Private.Windows.Core";
+    private const string SYSTEM_PRIVATE_WINDOWS_GDIPLUS = "System.Private.Windows.GdiPlus";
+    private const string SYSTEM_WINDOWS_FORMS = "System.Windows.Forms";
+    private const string SYSTEM_WINDOWS_FORMS_ANALYZERS = "System.Windows.Forms.Analyzers";
+    private const string SYSTEM_WINDOWS_FORMS_ANALYZERS_CSHARP = "System.Windows.Forms.Analyzers.CSharp";
+    private const string SYSTEM_WINDOWS_FORMS_ANALYZERS_VISUALBASIC = "System.Windows.Forms.Analyzers.VisualBasic";
+    private const string SYSTEM_WINDOWS_FORMS_ANALYZERS_CODEFIXES_CSHARP = "System.Windows.Forms.Analyzers.CodeFixes.CSharp";
+    private const string SYSTEM_WINDOWS_FORMS_ANALYZERS_CODEFIXES_VISUALBASIC = "System.Windows.Forms.Analyzers.CodeFixes.VisualBasic";
+    private const string SYSTEM_WINDOWS_FORMS_DESIGN = "System.Windows.Forms.Design";
+    private const string SYSTEM_WINDOWS_FORMS_PRIMITIVES = "System.Windows.Forms.Primitives";
+    private const string SYSTEM_WINDOWS_FORMS_PRIVATESOURCEGENERATORS = "System.Windows.Forms.PrivateSourceGenerators";
+
+    private readonly string[] s_preCheckItems =
+    [
+        ACCESSIBILITY,
+        MICROSOFT_VISUALBASIC,
+        MICROSOFT_VISUALBASIC_FACADE,
+        MICROSOFT_VISUALBASIC_FORMS,
+        MICROSOFT_PRIVATE_WINFORMS,
+        SYSTEM_DESIGN_FACADE,
+        SYSTEM_DRAWING_COMMON,
+        SYSTEM_DRAWING_DESIGN_FACADE,
+        SYSTEM_DRAWING_FACADE,
+        SYSTEM_PRIVATE_WINDOWS_CORE,
+        SYSTEM_PRIVATE_WINDOWS_GDIPLUS,
+        SYSTEM_WINDOWS_FORMS,
+        SYSTEM_WINDOWS_FORMS_ANALYZERS,
+        SYSTEM_WINDOWS_FORMS_ANALYZERS_CSHARP,
+        SYSTEM_WINDOWS_FORMS_ANALYZERS_VISUALBASIC,
+        SYSTEM_WINDOWS_FORMS_ANALYZERS_CODEFIXES_CSHARP,
+        SYSTEM_WINDOWS_FORMS_ANALYZERS_CODEFIXES_VISUALBASIC,
+        SYSTEM_WINDOWS_FORMS_DESIGN,
+        SYSTEM_WINDOWS_FORMS_PRIMITIVES,
+        SYSTEM_WINDOWS_FORMS_PRIVATESOURCEGENERATORS
+    ];
+
+    /// <summary>
+    ///  Raised whenever the set of available/checked assemblies changes
+    ///  (e.g. after picking a new source folder or target framework), so a
+    ///  host can enable/disable its own action controls accordingly.
+    /// </summary>
+    public event EventHandler? AvailabilityChanged;
+
+    public AssetSelectionControl()
+    {
+        InitializeComponent();
+
+        _pathToArtefactsRepoTextBox.TextChanged +=
+            (sender, e) => DeployAvailableRuntimes();
+
+        _availableDesktopRuntimesComboBox.SelectedIndexChanged +=
+            (sender, e) => DeployAvailableAssemblies();
+
+        _controlsForEnablingHandling =
+        [
+            _availableDesktopRuntimesComboBox,
+            _checkForRespectiveRefAssembliesCheckBox,
+            _availableAssembliesListView,
+        ];
+    }
+
+    public AssetSelectionControl(
+        RuntimeDeploySettingsService settings,
+        RuntimeDeployStatusService statusService,
+        ILogger<AssetSelectionControl> logger) : this()
+    {
+        _settings = settings;
+        _statusService = statusService;
+        _logger = logger;
+    }
+
+    protected override void OnLoad(EventArgs e)
+    {
+        base.OnLoad(e);
+        _pathToArtefactsRepoTextBox.Text = _settings?.SourceArtefactsFolder ?? string.Empty;
+    }
+
+    /// <summary>
+    ///  Re-reads the source folder from settings and, if unchanged, refreshes
+    ///  the assembly list (e.g. because the exclusion list changed).
+    /// </summary>
+    public void RefreshFromSettings()
+    {
+        string sourceArtefactsFolder = _settings?.SourceArtefactsFolder ?? string.Empty;
+        if (!string.Equals(_pathToArtefactsRepoTextBox.Text, sourceArtefactsFolder, StringComparison.OrdinalIgnoreCase))
+        {
+            _pathToArtefactsRepoTextBox.Text = sourceArtefactsFolder;
+            return;
+        }
+
+        DeployAvailableAssemblies();
+    }
+
+    /// <summary>Whether any assemblies are currently listed (a source folder + target were picked).</summary>
+    public bool HasAssemblies => _availableAssembliesListView.Items.Count > 0;
+
+    /// <summary>The currently checked assemblies.</summary>
+    public DesktopAssemblyInfo[] GetCheckedAssemblies()
+        =>
+        [
+            .. from ListViewItem item in _availableAssembliesListView.Items
+               where item.Checked
+               select (DesktopAssemblyInfo)item.Tag!
+        ];
+
+    /// <summary>The currently selected source target's TFM/configuration item, if any.</summary>
+    public TargetFrameworkSourceItem? SelectedSourceTarget
+        => _availableDesktopRuntimesComboBox.SelectedItem as TargetFrameworkSourceItem;
+
+    /// <summary>
+    ///  Finds the first listed assembly (checked or not) that has at least
+    ///  one ref-assembly file, used to locate the source ref-assembly base
+    ///  path for a copy/package operation.
+    /// </summary>
+    public DesktopAssemblyInfo? FindFirstWithRefAssemblies()
+        => (from ListViewItem item in _availableAssembliesListView.Items
+            let assemblyInfo = (DesktopAssemblyInfo)item.Tag!
+            where assemblyInfo.RefAssemblyFiles is not null
+                && assemblyInfo.RefAssemblyFiles.Length > 0
+            select assemblyInfo)
+           .FirstOrDefault();
+
+    private void HandleControlEnabling(bool enable, params Control[] excludeControlsForHandling)
+    {
+        foreach (var control in _controlsForEnablingHandling)
+        {
+            if (!excludeControlsForHandling.Contains(control))
+            {
+                control.Enabled = enable;
+            }
+        }
+    }
+
+    private void DeployAvailableRuntimes()
+    {
+        if (string.IsNullOrWhiteSpace(_pathToArtefactsRepoTextBox.Text))
+        {
+            _availableDesktopRuntimesComboBox.Items.Clear();
+            _availableAssembliesListView.Items.Clear();
+            HandleControlEnabling(false);
+            AvailabilityChanged?.Invoke(this, EventArgs.Empty);
+        }
+        else
+        {
+            try
+            {
+                _availableDesktopRuntimesComboBox.Items.Clear();
+                _availableAssembliesListView.Items.Clear();
+                _gitHubRepoManager = new(_pathToArtefactsRepoTextBox.Text);
+
+                var targets = _gitHubRepoManager
+                    .GetAvailableTargets();
+
+                _availableDesktopRuntimesComboBox.Items.AddRange(targets);
+                if (targets.Length > 0)
+                {
+                    _availableDesktopRuntimesComboBox.SelectedIndex = targets.Length - 1;
+                }
+
+                HandleControlEnabling(true);
+            }
+            catch (Exception ex)
+            {
+                _logger?.LogError(ex, "Could not enumerate RuntimeDeploy source targets.");
+                _statusService?.ReportException(ex);
+                HandleControlEnabling(false, _availableDesktopRuntimesComboBox);
+            }
+            finally
+            {
+                AvailabilityChanged?.Invoke(this, EventArgs.Empty);
+            }
+        }
+    }
+
+    private void DeployAvailableAssemblies()
+    {
+        if (_availableDesktopRuntimesComboBox.SelectedItem is null ||
+            _gitHubRepoManager is null)
+        {
+            return;
+        }
+
+        HashSet<string> excludedAssemblyNames = _settings?.GetExcludedAssemblyNames() ?? [];
+
+        var assemblies = _gitHubRepoManager.GetWinFormsRuntimeAssemblies(
+            (TargetFrameworkSourceItem)_availableDesktopRuntimesComboBox.SelectedItem,
+            _checkForRespectiveRefAssembliesCheckBox.Checked)
+            .Where(assembly => !excludedAssemblyNames.Contains(assembly.Name))
+            .ToArray();
+
+        _availableAssembliesListView.ConfigureDetailsView(checkBoxes: true);
+
+        _availableAssembliesListView.AddItemsWithColumnHeadersFromType(
+            assemblies,
+            addSourceDataToTag: true,
+            (nameof(DesktopAssemblyInfo.Name), "Assembly name"),
+            (nameof(DesktopAssemblyInfo.Path), "Path"));
+
+        _availableAssembliesListView.CheckItemsInFirstColumn(s_preCheckItems);
+
+        AvailabilityChanged?.Invoke(this, EventArgs.Empty);
+    }
+
+    private void PickPathToArtefactsButton_Click(object sender, EventArgs e)
+    {
+        FolderBrowserDialog browserDialog = new()
+        {
+            Description = "Pick the path to the WinForms artifacts folder:"
+        };
+
+        DialogResult dialogResult = browserDialog.ShowDialog();
+        if (dialogResult == DialogResult.OK)
+        {
+            _pathToArtefactsRepoTextBox.Text = browserDialog.SelectedPath;
+            if (_settings is not null)
+            {
+                _settings.SourceArtefactsFolder = browserDialog.SelectedPath;
+            }
+        }
+    }
+}
